[IMP] pos_preparation_display: remove click order on last stage
In this commit, we remove the stage change when clicking on an order
in the last stage of a preparation display. We also remove the stage
change when all orderlines of an orders are set to todo=false in the
last stage. We finally put all orderlines as todo=false by default when
arriving in the last stage of a preparation display. This is because
the flow was a bit confusing for the user in this stage. The user was
not expecting the order to come back to the first stage when clicking
on it. He will now have to click on one of the two button below the
order to change its state.

Part-of: odoo/enterprise#49200