[FIX] stock_barcode(_mrp): scan tracked product
When a move line is for a tracked product, if its `qty_done` is equal or
greater than its `product_uom_qty`, the line will not be selected when
the product will be scanned.
It's annoying when this line has no lot yet.

How to reproduce:

- Create a product with a BoM kit and at least one tracked component
  (both product should have a barcode of course);
- In the Barcode App, create a new receipt;
- Scan the product kit and validate -> Before the validation, the kit is
  decomposed and a line for each component should be created;
- Scan the tracked component -> Instead of selecting the existing line,
  a new one is created. That means the user is not able to set the lot
  for this line without select it manually.

To fix that, the line will be selected even if it has enough `qty_done`
only if the product is tracked and there is no lot/serial number.

closes odoo/enterprise#50582

X-original-commit: cc21ca18d27f599fbcefa9e4c0b1d067b26ee04f
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Steve Van Essche <svs@odoo.com>