[IMP] knowledge: custom SQL query for get_user_sorted_articles
Implement a custom SQL query to drastically improve search performances on name
in Knowledge.

Increase the command debounce timing for the search query to apply to `500ms`,
because for slow typers, the rpc result will replace what they are trying to
type with a previous value, and with a debounce value of `200ms`, this occurs
far too often, resulting in a poor experience.

Query count:

- `test_get_all_user_sorted_articles` (`get_user_sorted_articles`), (14 -> 5):
  - Since the ordering in done fully in SQL, we don't need to pass through
    compute methods like `_compute_is_user_favorite` or
    `_compute_favorite_count` anymore.
  - The `search` and the `read` are done in the same query instead of 2 separate
    ones.

task-3554068C

Part-of: odoo/enterprise#48945