[FIX] account_reports: don't call database when toggling hide_0_lines filter
https://github.com/odoo/enterprise/commit/4bf10c05ec9fa0aa794a02b163f2afd858ffa613#diff-9c1cf6319c0d789e8b5f51f2f340e63f9c3ca54e3df3b9d6babc4c35670880c7 introduced the new hide_0_lines filter, allowing the user to hide the lines with an empty amount. This filter relies on a javascript treatment of the returned lines to decide whether or not to display them, but they are always all returned by the server.

The original implementation, however, made use of the toggleFilter() helper function, which modifies the options dict before calling the database to fully regenerate the report. In this case, this was useless, since the returned lines would be the same.

Instead of doing that, we now assign the option key without reloading everything, and then just recompute the visibility of the lines.

closes odoo/enterprise#49173

Signed-off-by: John Laterre (jol) <jol@odoo.com>