[REF] *: remove .oe_left and .oe_rigth
The first intention of this commit was to remove ´@extend´ but
it actually makes sense to remove ´.oe_left´ and ´.oe_right´ classes.

Since grid we can avoid "float" elements (see .oe_subtotal_footer)
and we only have to remove this class.
For the other case, we have to use ´.float-start' and ´float-end´
to replace ´oe_left´ and ´oe_right´.

closes odoo/enterprise#49211

Related: odoo/odoo#139199
Signed-off-by: Pierre Paridans (app) <app@odoo.com>