[FIX] web_studio: restrict search mock to expected call
This patch was designed to intercept a specific search call, not all of
them. The scope of the patch is now arbitrary limited to when no params
are given, which is how the expected intercept is calling it.

It still makes no sense to do it this way but at least the test pass.

closes odoo/enterprise#49324

Signed-off-by: Matthieu Stockbauer (tsm) <tsm@odoo.com>