[REF] account_inter_company_rules,l10n_mx_edi: refactoring of the Send & Print wizard (part 1)
Currently the Send & Print wizard is stored as regular model
and not a TransientModel.
This cause multiple problems including concurrency ones.
This PR switches the wizard back to a regular transient one.
We instead store the values of the wizard that needs to be
processed asynchronously in the related move.

closes odoo/enterprise#49310

Task-id: 3415101
Related: odoo/odoo#139311
Signed-off-by: Laurent Smet (las) <las@odoo.com>