[FIX] account_online_synchronization: Add client id in error subject
The commit(6b5eb6961bccd76c6e25948275499bece2f0075c) that creates
a support ticket for the user directly has the side effect that
user creates too much tickets for the same issue.
After discussion with support team and POs, we decide to add
the client id in the support ticket title to allow the support
finding duplicate tickets.
It's clearly not the best solution, but it will do the job in
a short term.

no task id

closes odoo/enterprise#49255

X-original-commit: 2d32d95059db2d331875a0cc3d96f4cf6911bc86
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>