[FIX] l10n_mx_edi: Handle invoice lines with zero quantities.
Prior to 5f4b3ef2c8f62357d736f32702968668787e8e4c, it was possible to
have SO lines with zero as quantity. This was broken with the
`valor_unitario` computation in `_l10n_mx_edi_get_invoice_cfdi_values`
because it was relying on those quantities.

This fix will simply exclude the lines with 0 quantities when generating
the cdfi values

closes odoo/enterprise#49674

X-original-commit: 2f8d1c5723b54242b3853eb0fa5b2f5fea84235e
Signed-off-by: Laurent Smet (las) <las@odoo.com>