[FIX] hr_payroll: fix contract form view permissions
Before this commit, when trying to open a contract with a user with the
permissions 'Payroll Manager' and 'Employee Manager', the button
``action_open_salary_attachments`` on the contract form view would cause
the error ``EvalError: Can not evaluate python expression:
(bool(salary_attachments_count == 0))`` because the button used the
field ``salary_attachments_count`` in its invisible attribute. However,
not all all users have the access rights to this field.

With this commit, the button's access rights are limited to those of
``salary_attachments_count``, namely ``group_hr_payroll_user``

closes odoo/enterprise#49209

Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>