[FIX] account_reports: fix tour undeterministic with sort_lines
The tour could fail at this step if the click on the line happened
before the sort.
You can reproduce it by adding a time.sleep(2) in the python
function sort_lines.
We just define the line that should be clicked so we're sure the
sort has happened.

Linked to runbot error 27633, 27819

closes odoo/enterprise#49410

X-original-commit: 4e6abfac8c9e9a847d3bf7d1e927391a2809da97
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Wala Gauthier (gawa) <gawa@odoo.com>