[FIX] web_studio: viewEditor: correctly compute fieldsInArch
A view editor's panel will present the fields that are not yet in the arch for a user
to add them in the view via DragAndDrop. It filters out the field in the arch from the fields of the model

Before this commit, the fields in the arch were not correctly computed, and may generate inconsistencies.

After this commit, this works correctly.

closes odoo/enterprise#49353

X-original-commit: 6ff14e1071e91edbb7ae3dc5f51234b5c8e0640c
Signed-off-by: Florent Dardenne (dafl) <dafl@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>