[IMP] pos_restaurant_appointment: start appointment ressource by floor
When we used to create appointments resources, we called them
"table name - floor name", the problem was that table names were often
a number. Some applications rely on resources (e.g. Appointments,
Scheduling) and because we start with a number, the list of resources
looks messy.

Now we've changed the name format of the resources created like this:
"floor name - table name".

closes odoo/enterprise#49398

X-original-commit: c9e50175d5f9fdf2400ce9c58c5a3222425e921c
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>
Signed-off-by: David Monnom (moda) <moda@odoo.com>