[FIX] planning: display employee avatar for user with no hr access
Steps
=====
- Install module planning
- Log in as an user with planning access but no hr access
- Open planning main gantt view

Issue
=====
Empty avatars are displayed for employees.

Cause
=====
The model used to display avatar of employees is hr.employee. But user
with no hr access have no read access to this model.

Fix
===
If the current user has no hr access, the model hr.employee.public is
used to display the avatar.

closes odoo/enterprise#49363

X-original-commit: 8e1b1fb33f87b80b6869e6846c4994141a44d7e6
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>