[FIX] currency_rate_live: change banxico token
MX users using Banxico provider may have the scheduled action
"Currency: rate update" to fail sometimes due to api key limit.

Because of this the rate for the day cannot be fetched and an old
one may be used, severely affecting users accounting.

As explained in the Banxico documentation[1], any token has a dual limit:
- 10000 consults per day
- 200 consults every 5 minutes

Usually The latter is the cause of the errors in the cron as we use the same
token across all versions and lots of users may still be using the default one
causing a lot of requests with the same token in the same timeframe

A temporary mitigation to the issue is to change the currently used
token in all versions with a new one

[1]https://www.banxico.org.mx/SieAPIRest/service/v1/doc/limiteConsultas

opw-3541164

closes odoo/enterprise#49369

X-original-commit: e41dc6aadd62950c652650d2de0e71cc11d9066a
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Andrea Grazioso (agr) <agr@odoo.com>