[FIX] l10n_br_avatax: add proper parameters for credit and debit notes
The buttons and triggers already are enabled on these, but we forgot
to specify some API parameters.

This specifies the operationType in the header, instead of on each
line. The operationType (standardSales, salesReturn or
amountComplementary) is determined based on the move_type and
debit_origin_id.

invoiceRefs is specified to refer to the original invoice and
description is required for both.

Finally, we need to invert the tax amounts for returns since Avatax
doesn't do that for us.

This is based on odoo/enterprise#49498 for Odoo 16 but includes debit
note support as well since l10n_br indirectly depends on
account_debit_note in Odoo 17. It also adds the default operation type
and invoice refs directly in the mixin to avoid having to create an
l10n_br_avatax_sale module like in Odoo 16.

closes odoo/enterprise#49513

Signed-off-by: Josse Colpaert <jco@odoo.com>