[FIX] hr_holidays_gantt: get all employee ids
Steps to reproduce:
-------------------
- go to Time-Off / Overview;
- remove default "My Team" filter;
- group by "Company";
- add a custom group by on "Employee".

Issue:
------
A traceback occurs.

Cause:
------
When we try to get leaves (via `leave_mapping`) using employee id,
the id doesn't exist in this dictionnary.

It is because we don't have employee to initialise `leave_mapping`.
In fact, the `tag_employee_rows` method doesn't return employee ids.

Solution:
---------
As the `tag_employee_rows` method is recursive via itself,
it is necessary to update the internal employee id set variable
with the employee ids of the subrows.

This method will return all the employees concerned by the rows
and we can build the dictionary with their leaves.

opw-3547369

closes odoo/enterprise#49553

X-original-commit: f2fe0fd881955467af3a1ca5147bf8cd02966e4e
Signed-off-by: Sofie Gvaladze (sgv) <sgv@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>