[IMP] l10n_at_reports: permanently unfold top level report lines
Currently all the lines of the reports are folded.
Thus basically all the report information is hidden;
The user first has to unfold some top level lines.

This goes for the top level line of the profit and loss statement
and the top 2 level lines of the balance sheet.

closes odoo/enterprise#47405

Related: odoo/odoo#135485
Related: odoo/documentation#5859
Signed-off-by: Josse Colpaert <jco@odoo.com>