[FIX] *_extract: fix inconsistent OCR status check
Depending of the type document, the logic that handles the background
OCR status check was different, but it shouldn't have been the case.

The logic that is implemented for invoices should be used for expenses
and resumes as well.

See commit fcb7fdf for the reasons behind the cursor commit.

closes odoo/enterprise#49458

X-original-commit: 88b916819d8818e4822f9084b2051b928fea127d
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>