[FIX] l10n_mx_edi_40: Allow get the SAT status on CFDI 3.3
Fixed the error that its present on CFDI 3.3 records when they are
cancelled, but was signed on CFDI 3.3, and the `l10n_mx_edi_40`
module is installed.

When you cancel a document, the attachment is removed from the EDI
Document, but Odoo already support to get the attachment from the
record, but only support documents signed in CFDI 4.0, and all the
documents cancelled and signed in 3.3 returns 'Not Found', now try to
get the attachment in any version.

closes odoo/enterprise#49432

closes odoo/enterprise#49794

X-original-commit: 069afc010e504bec615174c840fae2d174cd34b2
Signed-off-by: Habib Ayob (ayh) <ayh@odoo.com>