[IMP] account_online_synchronization: Add a connection_state_details field
The aim of this commit is adding a simple JSON field on account online link
model to handle a future flow where we save the last connection state details
(if the connection is in error for example) to know what we have to do
for the user. We need to implement this flow because the commit
(ed7d7486f30e076e6040a2d528a2c4926225bfa9) that adds the new
asynchronous flow (by using cron) doesn't handle all the error messages.
It means that the user won't see that he/she receives an error and
it won't trigger the redirect warning error that allows the user
to open a support ticket directly (with useful info for support team).
The feature is divided in 2 commits, one to add the freeze dependent
things (the new field) and another one to add all the logic.

task-3568712

closes odoo/enterprise#49447

Signed-off-by: William André (wan) <wan@odoo.com>