[FIX] account_report: by default hide lines
In pull request https://github.com/odoo/enterprise/pull/45291, we introduced a
new feature that allows you to apply a filter on a report, enabling the ability
to hide lines with zero values based on certain conditions.

With this enhancement, you can now configure the filter to have one of three
settings: "enabled by default," "optional," or "never."

Prior to this commit, the "enabled by default" setting was not functioning as
expected. This issue came from an initial check within the first if statement
that evaluated previous options. The problem was that when you initially entered
the report, the previous options dict were empty, and so the options is put to
false which was not the intended behavior.

closes odoo/enterprise#49451

Task: 3568441
Signed-off-by: John Laterre (jol) <jol@odoo.com>