[FIX] voip: properly display calls in “calling” and “ongoing” states
If for some reason (e.g. a power outage in the middle of a call) the
call is never properly terminated, it may be stuck in the “calling” or
“ongoing” state.

Prior to this commit, trying to display a call that was stuck in the
wrong state would cause a crash.

After this commit, the call will be displayed correctly, even if it is
in an improper state.

closes odoo/enterprise#49465

Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>