[IMP] sale_stock_renting: create pickings for rental orders
SALE_STOCK_RENTING:
It will be possible to create pickings from rental orders by activating a new setting
in the General Settings of Rental.
Works with:

1-2-3 steps of delivery/reception
serial numbers
rental dates: they will be used for the scheduled dates of the pickings
SALE_MRP_RENTING:
New module to correctly compute the qty_delivered and the qty_returned on rental order lines for kit products
as well as manage the rewriting of ordered quantities on rental orders.

see odoo/odoo#135425
task #3109952

Part-of: odoo/enterprise#47376