[IMP] knowledge: improve the UX of the cover dialog
In PR [1], the media dialog has been improved in order to make the UX
better.

This commit adapts the `KnowledgeCoverDialog`, in order to be compatible
with the new media dialog. More precisely, these adaptations are the
following:
- the dialog height is now fixed and has the same size as when it is
fully filled by images -> addition of the `h-100` class.
- there is now a small footer displaying either an arrow to scroll, the
"Load More" button or the "All images have been loaded" text. In order
for this button to be updated, we need a scroll listener on the modal
body and therefore we need the reference to the modal in
`KnowledgeCoverSelector` (which extends `ImageSelector` and so
`FileSelector`, which uses this ref)-> this commit adds this ref.

[1]: https://github.com/odoo/odoo/pull/135875

task-3441587

closes odoo/enterprise#49535

Related: odoo/odoo#135875
Signed-off-by: Romain Derie (rde) <rde@odoo.com>