[FIX] account_reports: pick the right company's coordinates when exporting with branches selected
When multiple branches are active in the company selector, if they correspond to the full branch hierarchy, we want to enable all the buttons that are not branch_allowed. This is done so that we prevent a user generating by mistake some legal export for only some branches, forgetting to select the rest, and then submitting it.

However, before this commit, we did not check the company used for export corresponded to the top of the hierarchy. It was hence possible to generate an export file with a different active company, and the engine would then consider the main company of the selector as the sender of the file, instead of the branches' root.

Part-of: odoo/enterprise#50190