[FIX] helpdesk_sale_timesheet: prevent ticket stat button traceback
When a user changes product type of a service type product
which is used in a sale order and then try to access helpdesk
ticket linked to that sale order, the error is generated.

Steps to reproduce:
- Install helpdesk_sale_timesheet module.
- Create a sale order by selecting a customer and adding product
'Customer Care (Prepaid Hours)' (which is a service type product) and
click on confirm button.
- Now create a helpdesk ticket in VIP SUPPORT and select the same customer
and sale order which you have created.
- Now again go to same sale order and click on 'Customer Care (Prepaid Hours)'
product and change its type to 'consumable' from 'service', a warning will occur
close it and click on save button.
- Now click on helpdesk ticket smart button, the error will occur.

Error: StopIteration: null

The error is occurring because we are getting the iterable has reached to
its end in next function when condition in it is getting False over here -
https://github.com/odoo/enterprise/blob/1188db8cba436b887331bef8e07f6f1fafbe1982/helpdesk_sale_timesheet/models/sale_order.py#L37

To solve this error we are giving default value to next function.

sentry-4560590481

closes odoo/enterprise#50023

X-original-commit: b5afbf4be6a1160d7c4ec8262ceaeb011800eec1
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>