[FIX] account_report: total line page break
When activating the "Total Below" section feature, the total line will adopt the
attributes of the section line. More over, in the report settings we have the
option to define a page break attribute for a line.

Now if the section's line possesses the page break attribute, this causes the
total line to also have it, which is not the desired behavior. This commit aims
to separate the "Total Below" sections from the "print_on_new_page" field.

closes odoo/enterprise#49544

Task: 3566903
Signed-off-by: John Laterre (jol) <jol@odoo.com>