[FIX] planning: make resource avatar visible for non hr user
In planning, avatar can be displayed for resource.resource, hr.employee
and hr.employee.public records. Avatar of resource.resource records is
a computed field that return the avatar of the hr.employee record linked
to the resource if any. The problem is that not all user of planning
have access to hr.employee records.

This was not a problem in versions <16.0 but since,
https://github.com/odoo/odoo/commit/85bd6e62a59fef3086a9ae2138209cc73f086227,
access rights are checked when accessing an image through the route /web/images/<model>/<id>/<field>.

This commit avoids empty avatar for employees by falling back on
hr.employee.public avatar if the current user has no access to
hr.employee model.

closes odoo/enterprise#50078

X-original-commit: 7d2929ad00c7f0381a9d2003909ae4ec0f4c984a
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>