[IMP] account_reports: allow exporting reports for vat sub-hierarchies in branches
With the new branches mechanism, a button can either be branch_allowed or not. If it is not (which is the default, applied nearly everywhere), it will never be displayed when the full branch hierarchy is not selected.

This was limitating on tax-related reports, since they should make it possible to submit the taxes of some branches separately from the rest of the hierarchy. For that, we use VAT numbers: branches with the same VAT number are considered as submitting their tax reports together. Therefore, we don't disable the buttons for them if the currently selected sub-hierarchy contains all the companies sharing the VAT number of the main company chosen in the company selector.

Part-of: odoo/enterprise#50190