[FIX] l10n_lu_reports: wrong sign in the formula
Now the default external values are created with the negative sign, as the amount on the accounts is in credit. We need to add a minus in front to correct the amount.
The relevant test was created for an `in_invoice`, which is why the default value was not tested correctly.

closes odoo/enterprise#49575

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>