[FIX] whatsapp: whatsapp channel should be deactivate after 24 hours
Steps:
1. Install the WhatsApp module.
2. Create one WhatsApp account.
3. Send a template message from the account to a recipient's number.
4. Reply to the message from the WhatsApp application.
5. A new channel will be created for this conversation.
6. After 24 hours, the channel will be blocked from sending messages from the
   channel
7. Now send another template to that user.
when sending the new template it will post a message in the channel and that
channel will be allowed to send messages.
This working is wrong because it should work like a deactivate channel after 24
hours of the last message of the recipient,

Problem:
_compute_whatsapp_channel_valid_until method uses using _channel_last_message_ids
 method to retrieve the last message so, in this case, the last message will be
the message that can be posted by any user but we want the recipient's last message.

Solution:
instead of using _channel_last_message_ids created a new method to retrieve the
last message of the customer in the channel.

closes odoo/enterprise#49686

X-original-commit: b1f093325afa6189254c1b93cb93274f82106437
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>