[FIX] website_sale_renting: fix renting dates on product page
Before this commit, some tools of website_sale_renting were broken due
to a loading problem of the overrides of the Variant mixin. For
instance, the price on the product page was not depending on the rental
period anymore. We now force them to be executed by prepending the js
files in the manifest.

Furthermore, the dates were not always given in the context. In this
case, the fallback is on the cart dates, if any.

task-3570178

closes odoo/enterprise#49610

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>