[FIX] account_reports: properly disable export buttons in case of multicompany
When generating the options with multiple companies enabled in the selector, we want to disable the non branch-allowed export buttons if a branch is not fully selected (to avoid people wrongly exporting only part of the data they want). However, this was done in a too restrictive way, disregarding the company selection operated in the options dictionary, and considering only one branch hierarchy at a time could be selected.

With this commit, we allow selecting multiple hierarchies, as long as all their sub-branches are selected. This way, people can perform consolidation properly on reports using the 'selector' filter_multi_company values.

Part-of: odoo/enterprise#50190