[FIX] approvals: allow deletion of draft approval
There is a business check on attachment unlink that needs to access
approval state. This is why the attachment needs to be unlinked before
the approval itself.

steps to reproduce:
- create a draft approval and attach a document
- delete the approval.

before this commit:
odoo.exceptions.MissingError: Record does not exist or has been deleted.
(Record: approval.request(8,), User: 2)

after this commit:
approval is deleted

opw-3558437

closes odoo/enterprise#51095

X-original-commit: 13033b51565dc73028bbb93bebc9ecf598a72a3b
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Nicolas Danhier (nda) <nda@odoo.com>