[FIX] account_reports: tax closing: correctly filter data in multi-company
When enabling multi-company on the tax reports (through tax units, branches, or by enabling the 'selector' filter_multicompany value), one tax closing entry is created per company considered by the report. However, the amounts reported in these were incorrect, as they essentially were being mixed between companies.

This bug was due to the get_options() call added in a recent commit. Because of it, the options got regenerated, and the forced mutli_company key was ignored (init_options_multicompany does not look at previous options ; it's recomputed at every call depending on the active companies). Because of that, the previously forced multi_company key had no effect anymore. This is the proper way to handle options. Keys like that should never be forced. We fix that by forcing the active companies when calling get_options.

Part-of: odoo/enterprise#50190