[FIX] account_reports: select correct company when doing tax closing with branches
When performing the tax closing on a company with branches, one move is created per branch, in a journal that is likely to belong to the parent company. Before this fix, though the content and journal of these moves were correct, they were all made in the root company of the hierarchy. This happened because company_id was not assigned any value when creating the closing move, so it was by default assigned to the company of the provided journal.

closes odoo/enterprise#50190

Related: odoo/odoo#141033
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>