[FIX] sale_timesheet_enterprise: change billable_time_target visibility
Before this commit, the field is only visible if the user has no access
to timesheet app. The problem is the field could be prefetched once
the user has access to timesheet app even if he is not `HR User`.

This commit changes the `groups` defined in definition of
`billable_time_target` field to only fetch the field if the user
is at least HR User, however, the field will be only visible in
the form view of `hr.employee` model if the user is at least Timesheet
Approver.

task-3573317

closes odoo/enterprise#49712

Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>