[FIX] account_reports: Wrong total on moves in Aged Payable
Current behavior before PR:

In the Aged Payable report, the total column display the totals for every line.
On move lines, the coluwn shows a "0" which does not make sense.

Desired behavior after PR is merged:

This commit makes it now so that there is no value displayed in that column for
moves but only for total lines like partner lines or totals.

closes odoo/enterprise#50086

X-original-commit: 940998cd13ed0f9773901d8499d251a330b191bf
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Thomas Becquevort (thbe) <thbe@odoo.com>